requirements.txt
langchain
langchain-groq
python-dotenv
streamlit
langchain-community #to use community integrations like pdfloader,RecursiveUrlLoader for RAG
langchain-huggingface
langchain-chroma
sentence-transformers

streamlit-mic-recorder  #to record live audio from the mic  #issue: streamlit_mic_recorder caches audio in session state and never clears it — so st.rerun() triggers the same input forever.


